chore: allow running a dedicated importer instance
